requirements.txt
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
appnope==0.1.3
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
build==1.2.1
CacheControl==0.14.0
certifi==2023.5.7
cffi==1.16.0
charset-normalizer==3.2.0
cleo==2.1.0
click==8.1.7
colorlover==0.3.0
comm==0.1.3
crashtest==0.4.1
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
distlib==0.3.8
dnspython==2.6.1
dulwich==0.21.7
email_validator==2.1.1
executing==1.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
fastjsonschema==2.19.0
filelock==3.14.0
fonttools==4.41.0
frozendict==2.3.8
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
inflection==0.5.1
installer==0.7.0
ipykernel==6.24.0
ipython==8.14.0
ipywidgets==8.1.1
jaraco.classes==3.4.0
jedi==0.18.2
Jinja2==3.1.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyterlab-widgets==3.0.9
keyring==24.3.1
kiwisolver==1.4.4
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mock==5.1.0
more-itertools==10.0.0
msgpack==1.0.8
multitasking==0.0.11
nbformat==5.9.2
nest-asyncio==1.5.6
numpy==2.2.1
orjson==3.10.3
packaging==23.1
pandas==2.2.3
parso==0.8.3
peewee==3.17.7
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pkginfo==1.10.0
platformdirs==4.3.6
plotly==5.18.0
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pyfredapi==0.9.1
Pygments==2.15.1
pyparsing==3.0.9
pyproject_hooks==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.9
pytz==2023.3
PyYAML==6.0.1
pyzmq==26.2.0
rapidfuzz==3.9.2
referencing==0.32.0
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.0
rpds-py==0.13.2
rprint==0.0.8
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.4.1
stack-data==0.6.2
starlette==0.37.2
ta==0.10.2
TA-Lib==0.4.32
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
tomlkit==0.12.5
tornado==6.3.2
traitlets==5.9.0
trove-classifiers==2024.5.22
typer==0.12.3
typing_extensions==4.12.2
tzdata==2023.3
ujson==5.10.0
urllib3==1.26.18
uvicorn==0.30.0
uvloop==0.19.0
virtualenv==20.26.2
watchfiles==0.22.0
wcwidth==0.2.6
webencodings==0.5.1
websockets==12.0
widgetsnbextension==4.0.9
xattr==1.1.0
yfinance==0.2.44
zipp==3.16.2
